[UPDATE] JWT 전환

+++ server_graphql/typedefs-resolvers/notepads.js
@@ -1,5 +1,8 @@
 const {gql} = require('apollo-server')
 const db = require('../models')
+const jwt = require('jsonwebtoken')
+
+const SECRET_KEY = process.env.SECRET_KEY
 
 const typeDefs = gql`
     type Notepad {
@@ -11,22 +14,33 @@ const typeDefs = gql`
     }
 `
 
+function tokenDecode(token) {
+    const decode = jwt.verify(token, SECRET_KEY);
+    if (!decode) {
+        return null;
+    }
+    return decode;
+}
+
 const resolvers = {
     Query: {
-        initCheck: async (parent, args) => {
-            console.log("초기화 : ", args.ID);
-            if (args.ID) {       // 사용자 데이터(Session)가 있다면
+        // TODO : ID 인자 필요없으면 지워라!
+        initCheck: async (parent, args, context) => {
+            const {req, res} = context;
+            const decode = tokenDecode(req.cookies.token);
+            if (decode === null) return res.status(401).text('Unauthorized');
+            if (decode.ID) {       // 사용자 데이터(Session)가 있다면
                 console.log("Session...OK");
             } else {                      // 세션이 없는데 Notepad 접근 시
                 return 'False';
             }
 
             const initUserSessionResult = await db.User_SESSION.findOne({
-                where: {user_id: args.ID}
+                where: {user_id: decode.ID}
             });
 
             const initNotepadResult = await db.Notepad.findAll({
-                where: {user_id: args.ID}
+                where: {user_id: decode.ID}
             })
 
             if (initUserSessionResult === null || initNotepadResult === null) {
@@ -48,61 +62,61 @@ const resolvers = {
             }
             return JSON.stringify(initData);
         },
+        // TODO : ERR_HTTP_HEADERS_SENT
         loadNotepad: async (parent, args, context) => {
             const {req, res} = context;
-            if (args.ID === null) {
-                return res.send({name: 'False'});
-            }
+            const decode = tokenDecode(req.cookies.token);
+            if (decode === null) return res.status(401).json({error: 'Unauthorized'});
             try {
                 const loadNotepadResult = await db.Notepad.findOne({
                     where: {name: args.name}
                 });
-                return res.send({
-                    name: loadNotepadResult.name,
-                    memo: loadNotepadResult.memo
-                });
+                return res.send(JSON.stringify(loadNotepadResult));
             } catch (err) {
-                return res.send({name: 'False'});
+                return res.send('False');
             }
         },
-        deleteNotepad: async(parent, args, context) => {
+        deleteNotepad: async (parent, args, context) => {
             const {req, res} = context
-            const deleteNotepadResult = await db.Notepad.findOne({
-                where: {name: args.name}
-            })
-            console.log(deleteNotepadResult.tab)
-            db.Notepad.destroy({
-                where: {name: args.name}
-            })
+            const decode = tokenDecode(req.cookies.token);
+            if (decode === null) return res.status(401).json({error: 'Unauthorized'});
 
-            db.User_SESSION.update({
+            await db.Notepad.findOne({
+                where: {name: args.name}
+            })
+            await db.Notepad.destroy({
+                where: {name: args.name}
+            })
+            await db.User_SESSION.update({
                 count: args.count,
                 active: 0
-            }, {where:{user_id: args.ID}});
+            }, {where: {user_id: decode.ID}});
 
             return res.send('OK');
         }
     },
     Mutation: {
         saveNotepad: async (parent, args, context) => {
-            console.log(args.ID);
             const {req, res} = context
+            const decode = tokenDecode(req.cookies.token);
+            if (decode === null) return res.status(401).json({error: 'Unauthorized'});
+
             if (args.name.indexOf('../') !== -1) {
                 return res.send('Unable to access.');
             }
-            if (args.ID === null) {
+            if (decode.ID === null) {
                 return res.send('False');
             }
 
             const USER_SESSION_DATA = {
-                user_id: args.ID,
+                user_id: decode.ID,
                 count: args.count,
                 active: args.activeIndex
             }
 
             console.log("유저 데이터 : ", USER_SESSION_DATA)
 
-            const userSessionResult = await db.User_SESSION.findOne({where: {user_id: args.ID}});
+            const userSessionResult = await db.User_SESSION.findOne({where: {user_id: decode.ID}});
             if (userSessionResult === null) {
                 db.User_SESSION.create({
                     user_id: USER_SESSION_DATA.user_id,
@@ -123,7 +137,7 @@ const resolvers = {
             }
 
             const NOTEPAD_DATA = {
-                user_id: args.ID,
+                user_id: decode.ID,
                 name: args.name,
                 memo: args.memo,
                 tab: args.activeIndex
